Added channels necessary to get timer to update GUI
This is an absolute mess that needs to be cleaned up. There are far too many channels I made
trying to figure this out. GTK objects are not thread safe so they can't be moved to seperate threads.
However it is still necessary to be able to update the GUI when multiple threads are running. In the
Glib library there is a MainContext::channel() API which provides the needed functionality to deal with
this and be almost identical to the standard library's message passing way of dealing with threads.